added

+++ Segment Tree/Coordinate Compression.cpp
@@ -1,13 +1,13 @@
 //In The Name of AllAH
 #include <bits/stdc++.h>
 using namespace std;
-#define ll long long
+#define int long long
 const int N = 5e4 + 7;
 int a[N], b[N], ans[N];
-int main() {
+int32_t main() {
 	ios_base::sync_with_stdio(0);
     cin.tie(0);
-    int n; cin >> n; ll a[n + 1];
+    int n; cin >> n; int a[n + 1];
     for(int i = 1; i <= n; i++) cin >> a[i];
     sort(a + 1, a + n + 1);
     int ind = 1;
